+++ client/web/src/auth/SignInPage.test.tsx
@@ -62,6 +62,7 @@ describe('SignInPage', () => {
             sourcegraphDotComMode?: SourcegraphContext['sourcegraphDotComMode']
             allowSignup?: SourcegraphContext['allowSignup']
             primaryLoginProvidersCount?: SourcegraphContext['primaryLoginProvidersCount']
+            authAccessRequest?: SourcegraphContext['authAccessRequest']
         }
     ) =>
         renderWithBrandedContext(
@@ -78,6 +79,7 @@ describe('SignInPage', () => {
                                 resetPasswordEnabled: true,
                                 xhrHeaders: {},
                                 primaryLoginProvidersCount: props.primaryLoginProvidersCount ?? 5,
+                                authAccessRequest: props.authAccessRequest,
                             }}
                             telemetryRecorder={noOpTelemetryRecorder}
                         />
@@ -226,6 +228,45 @@ describe('SignInPage', () => {
         } as AuthenticatedUser
 
         expect(render('/sign-in', { authenticatedUser: mockUser }).asFragment()).toMatchSnapshot()
+    })
+
+    it('renders redirect when there is only 1 auth provider', () => {
+        const withGitHubProvider: SourcegraphContext['authProviders'] = [
+            {
+                serviceType: 'github',
+                displayName: 'GitHub',
+                isBuiltin: false,
+                authenticationURL: 'http://localhost/.auth/gitlab/login?pc=f00bar&returnTo=%2Fsearch',
+                serviceID: 'https://github.com',
+                clientID: '1234',
+                noSignIn: false,
+            },
+        ]
+
+        expect(render('/sign-in', { authProviders: withGitHubProvider }).asFragment()).toMatchSnapshot()
+    })
+
+    it('does not render redirect when there is only 1 auth provider with request access enabled', () => {
+        const withGitHubProvider: SourcegraphContext['authProviders'] = [
+            {
+                serviceType: 'github',
+                displayName: 'GitHub',
+                isBuiltin: false,
+                authenticationURL: 'http://localhost/.auth/gitlab/login?pc=f00bar&returnTo=%2Fsearch',
+                serviceID: 'https://github.com',
+                clientID: '1234',
+                noSignIn: false,
+            },
+        ]
+
+        expect(
+            render('/sign-in', {
+                authProviders: withGitHubProvider,
+                authAccessRequest: { enabled: true },
+                allowSignup: false,
+                sourcegraphDotComMode: false,
+            }).asFragment()
+        ).toMatchSnapshot()
     })
 
     it('renders different prefix on provider buttons', () => {